possibly fixed all-auth no user signup adapter
