+++ config/settings/production.py
@@ -7,8 +7,6 @@ from sentry_sdk.integrations.celery import CeleryIntegration
 
 from sentry_sdk.integrations.redis import RedisIntegration
 
-from allauth.account.adapter import DefaultAccountAdapter
-
 from .base import *  # noqa
 from .base import env
 
@@ -183,21 +181,4 @@ LOGGING = {
 # ------------------------------------------------------------------------------
 
 # django-allauth
-## Disable new user registration. New user accounts must be created from within
-## Django Admin Panel by an existing administrator.
-## https://stackoverflow.com/questions/29794052/how-could-one-disable-new-account-creation-with-django-allauth-but-still-allow
-
-class NoNewUsersAccountAdapter(DefaultAccountAdapter):
-
-    def is_open_for_signup(self, request):
-        """
-        Checks whether or not the site is open for signups.
-
-        Next to simply returning True/False you can also intervene the
-        regular flow by raising an ImmediateHttpResponse
-
-        (Comment reproduced from the overridden method.)
-        """
-        return False
-
-ACCOUNT_ADAPTER = 'NoNewUsersAccountAdapter'
+ACCOUNT_ADAPTER = 'config.all_auth_adapters.NoNewUsersAccountAdapter'
